Updated Complicated Star
Working on a solution for drawing the even numbered complicated stars.

Current code draws one star with half the number of points, retraces the final stroke, and then bisects the previous angle to offset drawing the final half of the stars.  This solution might only work for six-sided stars, but more work is needed to see if a formula will work for all even-pointed stars.